refactor(ui): move help into profile menu, breathe out sidebar, round chat avatars
- Extract HelpButton into its own component and surface it as a "Get help"
  row inside the profile slide-out (above Sign out), matching the modern
  Linear/Notion/Slack pattern of putting support inside the account menu.
- Drop the lifesaver icon from the sidebar bottom row to reduce chrome.
- Bump horizontal padding throughout the sidebar (org switcher, New Chat,
  chat history, profile row) for more breathing room.
- Switch chat message avatars from rounded squares to circles.
